% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMpkg-package.R
\docType{package}
\name{PSMpkg-package}
\alias{PSMpkg}
\alias{PSMpkg-package}
\title{PSMpkg: Predictive Soil Mapping with R}
\description{
The description file is used to keep track
  of the packages used within this book.
}
\author{
\strong{Maintainer}: Tom Hengl \email{tom.hengl@envirometrix.net}

Other contributors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721) [contributor]
}

}
\keyword{internal}
